+++ src/error.rs
@@ -233,8 +233,6 @@ pub enum ParseRequestError {
 
     /// The request's multipart data was invalid.
     #[error("Invalid multipart data")]
-    #[cfg(feature = "multipart")]
-    #[cfg_attr(feature = "nightly", doc(cfg(feature = "multipart")))]
     InvalidMultipart(multer::Error),
 
     /// Missing "operators" part for multipart request.
@@ -262,7 +260,6 @@ pub enum ParseRequestError {
     UnsupportedBatch,
 }
 
-#[cfg(feature = "multipart")]
 impl From<multer::Error> for ParseRequestError {
     fn from(err: multer::Error) -> Self {
         match err {