+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.3] - 2021-03-22
 
-- Add `extension::OpenTelemetry` and delete `extension::Tracing`.
+### Added
+
+- Add `extension::OpenTelemetry`.
+
+### - [Changed]
+
+- Now all features are not activated by default.
+- Remove `multipart` feature.
 
 ## [2.6.2] - 2021-03-20